refactor: decompose the frontend part
