+++ client/src/components/InputUrl/InputUrl.module.scss
@@ -8,4 +8,9 @@
   align-items: center;
   width: 100%;
   background-color: #fff;
+  border-radius: 7px;
+}
+
+.shortner__helperText {
+  font-size: 16px;
 }
