+++ client/src/components/MainPage/MainPage.tsx
@@ -1,12 +1,12 @@
 import { Box } from "@mui/material";
 import InputUrl from "../InputUrl/InputUrl";
-import NameCurrency from "../NameCurrency/NameCurrency";
+import NameUrl from "../NameUrl/NameUrl";
 import styles from "./MainPage.module.scss";
 
 function MainPage() {
   return (
     <Box className={styles.shortner__selector}>
-      <NameCurrency>Your url:</NameCurrency>
+      <NameUrl>Your url:</NameUrl>
       <InputUrl />
     </Box>
   );
