+++ client/src/components/NameUrl/NameUrl.tsx
@@ -1,8 +1,8 @@
 import { Typography } from "@mui/material";
 import { IChildernProps } from "../../types/types";
-import styles from "./NameCurrency.module.scss";
+import styles from "./NameUrl.module.scss";
 
-function NameCurrency(props: IChildernProps) {
+function NameUrl(props: IChildernProps) {
   const { children } = props;
   return (
     <Typography variant="h6" component="h1" className={styles.text}>
@@ -11,4 +11,4 @@ function NameCurrency(props: IChildernProps) {
   );
 }
 
-export default NameCurrency;
+export default NameUrl;
