+++ client/src/components/ShortUrl/ShortUrl.module.scss
@@ -1,0 +1,22 @@
+.link_wrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 20px;
+  border: 2px solid #9db8f8;
+
+  @media screen and (max-width: 430px) {
+    flex-direction: column;
+    gap: 10px;
+  }
+}
+
+.link__textIcon {
+  display: flex;
+  gap: 10px;
+
+  @media screen and (max-width: 330px) {
+    flex-direction: column;
+    align-items: center;
+  }
+}
