+++ client/src/components/ShortnerBtn/ShortnerBtn.module.scss
@@ -6,19 +6,6 @@
 }
 
 .btn__cut {
+  display: flex;
   align-self: center;
-  display: flex;
-}
-
-.link_wrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px 20px;
-  border: 2px solid #9db8f8;
-}
-
-.link__textIcon {
-  display: flex;
-  gap: 10px;
 }
