+++ client/src/components/ShortnerBtn/ShortnerBtn.tsx
@@ -1,10 +1,10 @@
 import { Box, Button, Collapse, Paper, Typography } from "@mui/material";
 import { useState } from "react";
-import AddLinkIcon from "@mui/icons-material/AddLink";
 import useActions from "../../hooks/useActions";
 import useTypedSelector from "../../hooks/useTypedSelector";
 import { IShortnerBtn } from "../../types/types";
 import Loader from "../Loader/Loader";
+import ShortUrl from "../ShortUrl/ShortUrl";
 import ViewError from "../ViewError/ViewError";
 import styles from "./ShortnerBtn.module.scss";
 
@@ -13,14 +13,6 @@ function ShortnerBtn(props: IShortnerBtn) {
   const { link, error, loading } = useTypedSelector((state) => state.link);
   const [visible, setVisible] = useState(false);
   const fetchLink = useActions();
-
-  const [isCopied, setIsCopied] = useState(false);
-
-  const handleCopyClick = () => {
-    navigator.clipboard.writeText(link.shortUrl);
-    setIsCopied(true);
-    setTimeout(() => setIsCopied(false), 1000);
-  };
 
   const cutUrl = () => {
     fetchLink({ originalUrl: url });
@@ -43,24 +35,7 @@ function ShortnerBtn(props: IShortnerBtn) {
           <Loader />
         ) : (
           <Box>
-            {error ? (
-              <ViewError>{error}</ViewError>
-            ) : (
-              <Paper className={styles.link_wrapper}>
-                <Box className={styles.link__textIcon}>
-                  <AddLinkIcon />
-                  <Typography variant="body1">{link.shortUrl}</Typography>
-                </Box>
-                <Button
-                  onClick={handleCopyClick}
-                  disabled={isCopied}
-                  variant="contained"
-                  color="primary"
-                >
-                  {isCopied ? "Copied!" : "Copy"}
-                </Button>
-              </Paper>
-            )}
+            {error ? <ViewError>{error}</ViewError> : <ShortUrl link={link} />}
           </Box>
         )}
       </Collapse>
