+++ client/src/store/redusers/combineReducer.ts
@@ -4,4 +4,5 @@ import linkReducer from "./linkReducer";
 export const rootReducer = combineReducers({
   link: linkReducer,
 });
+
 export type RootState = ReturnType<typeof rootReducer>;
